+++ activity/brightpay/react/changelog.md
@@ -1870,3 +1870,14 @@
 
 ---
 
+# Brightpay Development - 2025-01-02
+
+**Commit #5**
+
+- Repository: brightpay/brightpay
+- Type: web-application
+- Language: react
+- Month: 1/2025
+
+---
+